// ui/src/utils/dashboard_client/api.ts
export interface LogsearchSearchTarget {
    /**
     * 
     * @type {string}
     * @memberof LogsearchSearchTarget
     */
    ip?: string;
    /**
     * 
     * @type {number}
     * @memberof LogsearchSearchTarget
     */
    kind?: number;
    /**
     * 
     * @type {number}
     * @memberof LogsearchSearchTarget
     */
    port?: number;
    /**
     * 
     * @type {number}
     * @memberof LogsearchSearchTarget
     */
    status_port?: number;
}
// ...
export interface LogsearchTaskModel {
    /**
     * 
     * @type {string}
     * @memberof LogsearchTaskModel
     */
    error?: string;
    /**
     * 
     * @type {number}
     * @memberof LogsearchTaskModel
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof LogsearchTaskModel
     */
    log_store_path?: string;
    /**
     * 
     * @type {LogsearchSearchTarget}
     * @memberof LogsearchTaskModel
     */
    search_target?: LogsearchSearchTarget;
    /**
     * 
     * @type {number}
     * @memberof LogsearchTaskModel
     */
    state?: number;
    /**
     * 
     * @type {number}
     * @memberof LogsearchTaskModel
     */
    task_group_id?: number;
}